requirements.txt
transformers[torch]
transformers
torch
torchvision
accelerate
xformers
streamlit
numpy 
opencv-python
Pillow